+++ GlobalSuppressions.cs
@@ -5,4 +5,4 @@
 
 using System.Diagnostics.CodeAnalysis;
 
-[assembly: SuppressMessage("CodeQuality", "IDE0051:Supprimer les membres privés non utilisés", Justification = "<En attente>", Scope = "member", Target = "~M:EnhancedMonsters.Patches.MenuManager_Patches.Start(MenuManager)")]
+[assembly: SuppressMessage("CodeQuality", "IDE0051")]